Add reusable navigation mock utility
