+++ Components/__tests__/DeleteButton.test.js
@@ -1,26 +1,12 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import { NavigationContainer } from '@react-navigation/native'; // Import NavigationContainer
+import { sleep } from '../../Test/Navigation';
 import DeleteButton from '../DeleteButton';
 
 const callback = () => { /* callback code */ };
 const label = 'LABEL';
 const navigate = 'NAV_TO';
-
-const sleepTime = 1;
-const mockedNavigate = jest.fn();
-
-jest.mock('@react-navigation/native', () => {
-  const actualNav = jest.requireActual('@react-navigation/native');
-  return {
-    ...actualNav,
-    useNavigation: () => ({
-      navigate: mockedNavigate,
-    }),
-  };
-});
-
-const sleep = (ms: number) => new Promise((resolve) => setTimeout(() => resolve(true), ms));
 
 describe('DeleteButton', () => {
   test('renders correctly with default settings', async () => {
@@ -35,7 +21,7 @@ describe('DeleteButton', () => {
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
     
-    await sleep(sleepTime);
+    await sleep();
   });
 
   test('renders correctly with textOnly = false', async () => {
@@ -51,7 +37,7 @@ describe('DeleteButton', () => {
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
 
-    await sleep(sleepTime);
+    await sleep();
   });
 
   test('renders correctly with textOnly = true', async () => {
@@ -67,6 +53,6 @@ describe('DeleteButton', () => {
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
     
-    await sleep(sleepTime);
+    await sleep();
   });
 });
