+++ Test/Navigation.js
@@ -1,0 +1,13 @@
+const mockedNavigate = jest.fn();
+
+jest.mock('@react-navigation/native', () => {
+  const actualNav = jest.requireActual('@react-navigation/native');
+  return {
+    ...actualNav,
+    useNavigation: () => ({
+      navigate: mockedNavigate,
+    }),
+  };
+});
+
+export const sleep = (ms = 1) => new Promise((resolve) => setTimeout(() => resolve(true), ms));
